Create secret auction code
This code is the secret auction code. This code considers the highest bidder as the winner of the secret auction. Firstly, it takes input from the user, it asks the  name of the bidder and then asks the bid and creates a bidding dictionary .If there are more bidders it takes the next input else the bidder with the highest bid is the winner.